+++ python_modules/dagster/dagster/_core/definitions/asset_check_factories/metadata_limit_checks.py
@@ -6,13 +6,11 @@ from dagster._core.definitions.events import AssetMaterialization
 from dagster._core.errors import DagsterInvalidDefinitionError
 from dagster._core.instance import DagsterInstance
 
-from ..asset_check_result import AssetCheckResult
 from ..asset_check_spec import AssetCheckKey, AssetCheckSeverity, AssetCheckSpec
 from ..asset_checks import AssetChecksDefinition
 from ..asset_key import AssetKey, CoercibleToAssetKey
 from ..assets import AssetsDefinition, SourceAsset
-from ..decorators.asset_check_decorator import multi_asset_check
-from .utils import assets_to_keys, unique_id_from_asset_keys
+from .utils import assets_to_keys, build_multi_asset_check_factory
 
 
 @experimental
@@ -97,27 +95,15 @@ def build_metadata_limit_checks(
             )
         return True, f"Value `{value}` is within range"
 
-    @multi_asset_check(
-        specs=[
+    return build_multi_asset_check_factory(
+        check_specs=[
             AssetCheckSpec(
-                f"{metadata_key.replace(' ','_')}_range_check",
+                f"{metadata_key.replace(' ','_')}_limit_check",
                 asset=asset_key,
                 description=description,
             )
             for asset_key in asset_keys
         ],
-        can_subset=True,
-        name=unique_id_from_asset_keys(list(asset_keys)),
+        check_fn=_result_for_check_key,
+        severity=severity,
     )
-    def _checks(context):
-        for asset_check_key in context.selected_asset_check_keys:
-            passed, description = _result_for_check_key(context.instance, asset_check_key)
-            yield AssetCheckResult(
-                passed=passed,
-                description=description,
-                severity=severity,
-                check_name=asset_check_key.name,
-                asset_key=asset_check_key.asset_key,
-            )
-
-    return [_checks]
